Basic backend start for blog
This is a lot but most of it is from package installs. I have installed:
- Express JS
- Mongoose
- Node JS
- MongoDB
I have basic endpoints for a user to be created updated and deleted. For now this just hits a local database on my own machine. I tested the endpoints with postman so they should be working as expected. This is a decent base and from here I will flesh out the user endpoints more and create other endpoints needed for a blog, Ill figure it out.