endpoint to login
Pretty simple, just adding the ability to login. However it is not secure. So Ill get to that after I make some other endpoints for blog posts and whatever I think about. I also want to do a small refactor of user.js or creating middlware instead of using console.log.

anywho, tested with postman and getting the right responses so ill merge to develop.