Basic JWT authentication implemented
When logging in, we now sign the request and we use the token generated from the response to verify update and delete actions. This auth process was moved into middleware so it can be expanded or used anywhere we think we might need it.

This can be built upon but its a good start and will allow me to get some front end work started with flutter. I tested this with postman and its looking pretty good so far.

Also installed prettier so expect some formatting updates as well with this commit.